backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
pdfplumber
python-docx
python-dotenv 